Update: Merge Dev Branch, Feature #4 Allow uppercase/lowercase letters and spaces,  etc
